Revert "refactor: Change `MutationType` effect/probability and `Cell` fitness member types from `double` to `float`."
This reverts commit 313aa370f970d197c567bc812dadf0c115b561fa.

+++ CellEvoX/include/ecs/Run.hpp
@@ -32,7 +32,7 @@ class Run {
   int positive_mutations = 0;
   int neutral_mutations = 0;
   int negative_mutations = 0;
-  float average_mutations = 0.0;
+  double average_mutations = 0.0;
   size_t total_cell_memory_usage = 0;
   size_t total_mutations_memory = 0;
   size_t total_graveyard_memory = 0;
